+++ Engine/spcOpenGL/gl4/pipeline/forward/gl4forwardpointlightrenderer.cc
@@ -119,7 +119,7 @@ GL4RenderTargetCube* GL4ForwardPointLightRenderer::GetPointLightShadowMap(Size i
   if (m_pointLightShadowMap.size() <= idx)
   {
     iRenderTargetCube::Descriptor desc{};
-    desc.Size = m_pointLightShadowMapSize;
+    desc.Size = static_cast<UInt16>(m_pointLightShadowMapSize);
     GL4RenderTargetCube* cubeMap = static_cast<GL4RenderTargetCube*>(m_device->CreateRenderTarget(desc));
